Adapt to newest libdatadog changes
