+++ components-rs/crashtracker.h
@@ -58,6 +58,34 @@ extern "C" {
 #endif // __cplusplus
 
 /**
+ * Disables the crashtracker.
+ * Note that this does not restore the old signal handlers, but rather turns crash-tracking into a
+ * no-op, and then chains the old handlers.  This means that handlers registered after the
+ * crashtracker will continue to work as expected.
+ *
+ * # Preconditions
+ *   None
+ * # Safety
+ *   None
+ * # Atomicity
+ *   This function is atomic and idempotent.  Calling it multiple times is allowed.
+ */
+DDOG_CHECK_RETURN struct ddog_VoidResult ddog_crasht_disable(void);
+
+/**
+ * Enables the crashtracker, if it had been previously disabled.
+ * If crashtracking has not been initialized, this function will have no effect.
+ *
+ * # Preconditions
+ *   None
+ * # Safety
+ *   None
+ * # Atomicity
+ *   This function is atomic and idempotent.  Calling it multiple times is allowed.
+ */
+DDOG_CHECK_RETURN struct ddog_VoidResult ddog_crasht_enable(void);
+
+/**
  * Reinitialize the crash-tracking infrastructure after a fork.
  * This should be one of the first things done after a fork, to minimize the
  * chance that a crash occurs between the fork, and this call.
@@ -97,6 +125,24 @@ DDOG_CHECK_RETURN
 struct ddog_VoidResult ddog_crasht_init(struct ddog_crasht_Config config,
                                         struct ddog_crasht_ReceiverConfig receiver_config,
                                         struct ddog_crasht_Metadata metadata);
+
+/**
+ * Reconfigure the crashtracker and re-enables it.
+ * If the crashtracker has not been initialized, this function will have no effect.
+ *
+ * # Preconditions
+ *   None.
+ * # Safety
+ *   Crash-tracking functions are not reentrant.
+ *   No other crash-handler functions should be called concurrently.
+ * # Atomicity
+ *   This function is not atomic. A crash during its execution may lead to
+ *   unexpected crash-handling behaviour.
+ */
+DDOG_CHECK_RETURN
+struct ddog_VoidResult ddog_crasht_reconfigure(struct ddog_crasht_Config config,
+                                               struct ddog_crasht_ReceiverConfig receiver_config,
+                                               struct ddog_crasht_Metadata metadata);
 
 /**
  * Initialize the crash-tracking infrastructure without launching the receiver.
@@ -364,6 +410,14 @@ HRESULT OutOfProcessExceptionEventCallback(const void *context,
                                            uint32_t *_size,
                                            uint32_t *_signature_count);
 #endif
+
+/**
+ * # Safety
+ * The `crash_info` can be null, but if non-null it must point to a Builder made by this module,
+ * which has not previously been dropped.
+ */
+DDOG_CHECK_RETURN
+struct ddog_VoidResult ddog_crasht_CrashInfo_demangle_names(struct ddog_crasht_Handle_CrashInfo *crash_info);
 
 /**
  * # Safety
